update navigation button
