+++ src/Categories.js
@@ -1,41 +1,64 @@
-import React, { useState, useEffect } from "react";
-import logo from "./logo.svg";
+import React from "react";
+import Carousel from "react-elastic-carousel";
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link,
+  BrowserRouter,
+} from "react-router-dom";
+import Images from "./images/Image";
 import "./App.css";
+import Sport from "./Sport";
+import { useHistory } from "react-router-dom";
 
-export default function Categories() {
-  const [ActivityDet, setADetails] = useState("");
-  const [Activity, setActivity] = useState("");
-  const [Type, setType] = useState("");
-  const [Participant, setParticipant] = useState("");
+const breakPoints = [
+  { width: 1, itemsToShow: 1 },
+  { width: 550, itemsToShow: 2 },
+  { width: 768, itemsToShow: 3 },
+  { width: 1200, itemsToShow: 4 },
+];
 
-  // fetch api //
-  useEffect(() => {
-    fetch("https://www.boredapi.com/api/activity")
-      .then((response) => response.json())
-      .then((data) => setADetails(data));
-  }, []);
-
+function Categories() {
+  const history = useHistory();
   return (
-    <div id="bored-container">
-      <h1>Bored?</h1>
-      <div id="box-container">
-        <button class="scale" id="box-button">
-          <img src={logo} className="Logo" alt="logo" />
-          <p>Type: {ActivityDet.type}</p>
-        </button>
-        <button class="scale" id="box-button">
-          <img src={logo} className="Logo" alt="logo" />
-          <p>Type: {ActivityDet.type}</p>
-        </button>
-        <button class="scale" id="box-button">
-          <img src={logo} className="Logo" alt="logo" />
-          <p>Type: {ActivityDet.type}</p>
-        </button>
-        <button class="scale" id="box-button">
-          <img src={logo} className="Logo" alt="logo" />
-          <p>Type: {ActivityDet.type}</p>
-        </button>
+    <BrowserRouter>
+      <Route path="/sport" component={Sport} />
+      <h1 style={{ textAlign: "center" }}>Categories</h1>
+      <div className="Slider">
+        <Carousel breakPoints={breakPoints}>
+            <button
+              class="scale"
+              id="box-button"
+              onClick={() => history.push("/sport")}
+            >
+              <img src={Images.recreation} className="Logo" alt="" />
+              <p>Recreational</p>
+            </button>
+          <button class="scale" id="box-button">
+            <img src={Images.education} className="Logo" alt="" />
+            <p>Education</p>
+          </button>
+          <button class="scale" id="box-button">
+            <img src={Images.cooking} className="Logo" alt="" />
+            <p>Cooking</p>
+          </button>
+          <button class="scale" id="box-button">
+            <img src={Images.charity} className="Logo" alt="" />
+            <p>Charity</p>
+          </button>
+          <button class="scale" id="box-button">
+            <img src={Images.sport} className="Logo" alt="" />
+            <p>Sport</p>
+          </button>
+          <button class="scale" id="box-button">
+            <img src={Images.social} className="Logo" alt="" />
+            <p>Social</p>
+          </button>
+        </Carousel>
       </div>
-    </div>
+    </BrowserRouter>
   );
 }
+
+export default Categories;
