+++ src/Sport.js
@@ -1,0 +1,11 @@
+import React from "react";
+import "./App.css";
+import { useHistory } from "react-router-dom";
+
+function Sport(){
+  return (
+      <h1>Sport page</h1>
+  );
+};
+
+export default Sport;
